% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNhold.R
\docType{data}
\name{BNhold}
\alias{BNhold}
\title{AID348 hold-out data using Burden Numbers for testing the EPX package}
\format{
A dataframe with 3946 rows and 25 variables:
\describe{
\item{WBN}{Burden numbers descriptor set with 24 variables.}
\item{y}{The response variable where 1 denotes active and 0 inactive.}
}
}
\usage{
BNhold
}
\description{
AID348 hold-out data with 24 burden numbers as explanatory variables.
Demonstrates in a timely manner \code{\link{epx}}, the phalanx-formation
algorithm in \pkg{EPX} and associated functions \code{\link{summary.epx}},
\code{\link{predict.epx}} , \code{\link{plot.epx}}, \code{\link{hit.curve}}.
}
\references{
Tomal, J. H., Welch, W. J., & Zamar, R. H. (2015).
Ensembling classification models based on phalanxes of variables with
applications in drug discovery.
\emph{The Annals of Applied Statistics},
\emph{9}(1), 69-93.
\url{http://doi.org/10.1214/14-AOAS778}
}
\keyword{datasets}
