% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/harvest.R
\docType{data}
\name{harvest}
\alias{harvest}
\title{Simulated dataset for testing the EPX package}
\format{
A dataframe with 190 rows and 4 variables:
\describe{
\item{LogP}{Octanol/water partition coefficient (-2 to 7).}
\item{MeltPt}{Melting point (120 to 280 degrees Celsius).}
\item{MolWt}{Molecular weight (200 to 800).}
\item{y}{The response variable where 1 denotes active and 0 inactive.}
}
}
\usage{
harvest
}
\description{
A simulated dataset from Yuan et al. (2012) with three explanatory variables.
Demonstrates in a timely manner \code{\link{epx}}, the phalanx-formation
algorithm in \pkg{EPX} and associated functions \code{\link{summary.epx}},
\code{\link{predict.epx}} , \code{\link{plot.epx}}, \code{\link{cv.epx}},
\code{\link{hit.curve}}.
}
\references{
Yuan, Y., Chipman, H. A., & Welch, W. J. (2012).
Harvesting Classification Trees for Drug Discovery.
\emph{Journal of Chemical Information and Modeling},
\emph{52}(12), 3169-3180.
\doi{10.1021/ci3000216}
}
\keyword{datasets}
